+++ webpack.config.js
@@ -1,0 +1,17 @@
+const path = require(`path`);
+
+module.exports = {
+  mode: `development`, // Режим сборки
+  entry: `./src/main.js`, // Точка входа приложения
+  output: { // Настройка выходного файла
+    filename: `bundle.js`,
+    path: path.join(__dirname, `public`)
+  },
+  devtool: `source-map`, // Подключаем sourcemaps
+  devServer: {
+    contentBase: path.join(__dirname, `./public`),
+    publicPath: 'http://localhost:8080/',
+    compress: true,
+    watchContentBase: true
+  }
+};
